[GTK][WPE] EWS should not wipe the JHBuild in the unapply patch step https://bugs.webkit.org/show_bug.cgi?id=206061
Reviewed by Aakash Jain.

The current unapply patch step on the EWS calls the script clean-webkit, which wipes everything.
Rebuilding the JHBuild takes around 30 minutes, even with ccache.
And its not needed to wipe the JHBuild directories on the clean-webkit step, because the WebKit
tooling for building the JHBuild (update-webkitgtk-libs) already detects when the moduleset has
been modified and wipes it when necessary.

This patch adds the optional argument --keep-jhbuild-directory to the script clean-webkit,
that the EWS will pass when doing the unapply patch step for platform GTK or WPE.

* BuildSlaveSupport/ews-build/steps.py:
(CleanWorkingDirectory.start):
* Scripts/clean-webkit:
(main):
* Scripts/webkitpy/common/checkout/scm/scm.py:
(SCM.discard_untracked_files):
* Scripts/webkitpy/common/checkout/scm/scm_unittest.py:

Canonical link: https://commits.webkit.org/219197@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@254359 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Tools/Scripts/webkitpy/common/checkout/scm/scm.py
@@ -213,9 +213,11 @@ class SCM:
     def untracked_files(self, include_ignored_files=False):
         self._subclass_must_implement()
 
-    def discard_untracked_files(self, discard_ignored_files=False):
+    def discard_untracked_files(self, discard_ignored_files=False, keep_webkitbuild_directory=False):
         for filename in self.untracked_files(discard_ignored_files):
             if self._filesystem.isdir(filename):
+                if keep_webkitbuild_directory and filename == "WebKitBuild":
+                    continue
                 self._filesystem.rmtree(filename)
             else:
                 self._filesystem.remove(filename)
